Time: 537 ms (34.77%), Space: 243.6 MB (16.04%) - LeetHub

+++ 0493-reverse-pairs/0493-reverse-pairs.cpp
@@ -30,7 +30,7 @@ public:
         int right=mid+1;
         int cnt=0;
         for(int i=low;i<=mid;i++){
-            while (right <= high && (long long)arr[i] > 2LL * arr[right])
+            while (right <= high && (long long)arr[i] > (long long)2* arr[right])
             right++;
             cnt+=right-(mid+1);
         }
